Rename: 커뮤니티 dto 위치 변경
커뮤니티 기능에서 dto는 공통으로 사용되므로 공통 패키지로 이동하였다. 그에 따른 import도 수정함

Related to: #37
